combine download_manager to utils
